# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Click Proposta v1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/complete-register": {
      "patch": {
        "operationId": "completeRegister",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 11
                  },
                  "cpf": {
                    "type": "string",
                    "minLength": 11,
                    "maxLength": 11
                  }
                },
                "required": [
                  "phone",
                  "cpf"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "406": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/service": {
      "post": {
        "operationId": "createService",
        "tags": [
          "Services"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "default": 201,
                      "type": "number"
                    }
                  },
                  "required": [
                    "message",
                    "statusCode"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "default": 500,
                      "type": "number"
                    }
                  },
                  "required": [
                    "message",
                    "statusCode"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/services/{idService}": {
      "delete": {
        "operationId": "deleteService",
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            },
            "in": "path",
            "name": "idService",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "default": 200,
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "default": 500,
                      "type": "number"
                    }
                  },
                  "required": [
                    "message",
                    "statusCode"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateService",
        "tags": [
          "Services"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            },
            "in": "path",
            "name": "idService",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "default": 200,
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "default": 500,
                      "type": "number"
                    }
                  },
                  "required": [
                    "message",
                    "statusCode"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/proposal/{id}": {
      "get": {
        "operationId": "getProposalById",
        "tags": [
          "Proposals"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                    },
                    "title": {
                      "type": "string"
                    },
                    "urlLogoImage": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "customersId": {
                      "type": "string"
                    },
                    "welcomeDescription": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "whyUs": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "challenge": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "solution": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "results": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "discount": {
                      "anyOf": [
                        {
                          "type": "number"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "totalPrice": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "userId": {
                      "type": "string"
                    },
                    "customerName": {
                      "type": "string"
                    },
                    "services": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "price": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "price",
                          "name"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "urlLogoImage",
                    "customersId",
                    "welcomeDescription",
                    "whyUs",
                    "challenge",
                    "solution",
                    "results",
                    "discount",
                    "totalPrice",
                    "userId",
                    "customerName",
                    "services"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}